Extracted common classes into own module
