+++ common/src/test/java/dk/dbc/solr/performance/LineSourceTest.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package dk.dbc.solr.performance.recorder;
+package dk.dbc.solr.performance;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
