+++ pom.xml
@@ -28,6 +28,7 @@
     </parent>
 
     <modules>
+        <module>common</module>
         <module>recorder</module>
     </modules>
     
@@ -46,45 +47,6 @@
             <build>
 
                 <plugins>
-                    
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-resources-plugin</artifactId>
-                        <executions>
-                            <execution>
-                                <id>shared-exec-java</id>
-                                <goals>
-                                    <goal>resources</goal>
-                                </goals>
-                                <phase>generate-sources</phase>
-                                <configuration>
-                                    <outputDirectory>${project.build.directory}/generated-sources/shared-exec</outputDirectory>
-                                    <resources>
-                                        <resource>
-                                            <directory>${basedir}/../shared-exec/src/main/java/</directory>
-                                            <filtering>true</filtering>
-                                        </resource>
-                                    </resources>
-                                </configuration>
-                            </execution>
-                        </executions>
-                    </plugin>
-                    
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-compiler-plugin</artifactId>
-                        <executions>
-                            <execution>
-                                <id>default-compile</id>
-                                <configuration>
-                                    <compileSourceRoots>
-                                        <compileSourceRoot>${basedir}/src/main/java</compileSourceRoot>
-                                        <compileSourceRoot>${project.build.directory}/generated-sources/shared-exec</compileSourceRoot>
-                                    </compileSourceRoots>
-                                </configuration>
-                            </execution>
-                        </executions>
-                    </plugin>                    
                     <plugin>
                         <groupId>org.apache.maven.plugins</groupId>
                         <artifactId>maven-dependency-plugin</artifactId>
@@ -137,6 +99,12 @@
             </build>
             
             <dependencies>
+                <dependency>
+                    <groupId>${project.groupId}</groupId>
+                    <artifactId>solr-performance-test-common</artifactId>
+                    <version>${project.version}</version>
+                </dependency>
+
                 <dependency>
                     <groupId>commons-cli</groupId>
                     <artifactId>commons-cli</artifactId>
